requirements.txt
appdirs==1.4.4
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.2
Django==3.2
django-autoslug==1.9.8
django-ckeditor==6.0.0
django-crispy-forms==1.11.2
django-filter==2.4.0
django-js-asset==1.2.2
django-multiselectfield==0.1.12
filelock==3.0.12
idna==2.10
mysqlclient==2.0.3
oauthlib==3.1.0
Pillow==8.2.0
pycodestyle==2.5.0
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
python-decouple==3.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
stripe==2.27.0
urllib3==1.26.4
virtualenv==20.4.7
